removed unwanted files

+++ app/Http/Controllers/ClientController.php
@@ -7,6 +7,8 @@ use App\Http\Requests\ClientUpdateRequest;
 use App\Models\Admin;
 use App\Models\Client;
 use App\Services\ClientService;
+use Exception;
+use Illuminate\Http\Response;
 use Illuminate\Routing\Controller as BaseController;
 use Illuminate\Support\Facades\Auth;
 
@@ -26,40 +28,51 @@ class ClientController extends BaseController
     {
         $user = Auth::user();
         $clients = $this->clientService->getAllClientsByUser($user->id);
-        
-        if($clients->count() == 0){
-            return response()->json("You don't have any clients yet");
 
+        if ($clients->count() == 0) {
+            return response()->json("You don't have any clients yet");
         }
         return response()->json($clients);
     }
 
     public function createClient(ClientCreateRequest $request)
-    {        
+    {
         $clientDetails = $request->validated();
         $clientDetails['profile_picture'] = $this->imageUpload($request->profile_picture);
-        $client = $this->clientService->storeClient(null,$clientDetails);
+        $client = $this->clientService->storeClient(null, $clientDetails);
 
         return response()->json($client);
     }
 
     public function updateClient(ClientUpdateRequest $request)
     {
-        
-        $clientDetails = $request->validated();
+        $user = Auth::user();
+        $client = $this->clientService->getClientById($request->id);
 
-        if ($request->hasFile('profile_picture')) {
-            $clientDetails['profile_picture']  = $this->imageUpload($request->profile_picture);
+        if ($user->can('update', $client)) {
+            $clientDetails = $request->validated();
+
+            if ($request->hasFile('profile_picture')) {
+                $clientDetails['profile_picture']  = $this->imageUpload($request->profile_picture);
+            }
+
+            $client = $this->clientService->storeClient($clientDetails['id'], $clientDetails);
+
+            return response()->json($client);
         }
-
-        $client = $this->clientService->storeClient($clientDetails['id'],$clientDetails);
-
-        return response()->json($client);
+        return response()->json(['Unauthorized'])->setStatusCode(Response::HTTP_UNAUTHORIZED);
     }
 
     public function deleteClient($id)
     {
-        return response()->json($this->clientService->deleteClient($id));
+        $user = Auth::user();
+        $client = $this->clientService->getClientById($id);
+
+        if ($user->can('delete', $client)) {
+            return response()->json($this->clientService->deleteClient($id));
+        }
+
+        return response()->json(['Unauthorized'])->setStatusCode(Response::HTTP_UNAUTHORIZED);
     }
 
     public function getClient($id)
